自动下载模型

+++ src/test/java/com/hankcs/hanlp/utility/TestUtility.java
@@ -21,19 +21,47 @@ import java.util.zip.ZipInputStream;
  */
 public class TestUtility
 {
+    static
+    {
+        ensureFullData();
+    }
+
+    public static void ensureFullData()
+    {
+        ensureData("data/model/crf", "http://nlp.hankcs.com/download.php?file=data", ".");
+    }
+
+    /**
+     * 保证 name 存在，不存在时自动下载解压
+     *
+     * @param name 路径
+     * @param url  下载地址
+     * @return name的绝对路径
+     */
     public static String ensureData(String name, String url)
     {
-        File target = new File(String.format("data/test/%s", name));
+        return ensureData(name, url, null);
+    }
+
+    /**
+     * 保证 name 存在，不存在时自动下载解压
+     *
+     * @param name 路径
+     * @param url  下载地址
+     * @return name的绝对路径
+     */
+    public static String ensureData(String name, String url, String parentPath)
+    {
+        File target = new File(name);
         if (target.exists()) return target.getAbsolutePath();
         try
         {
-            File testPath = new File("data/test");
-            if (!testPath.exists()) testPath.mkdirs();
-            String filePath = downloadFile(url, testPath.getAbsolutePath());
+            File parentFile = parentPath == null ? new File(name).getParentFile() : new File(parentPath);
+            if (!parentFile.exists()) parentFile.mkdirs();
+            String filePath = downloadFile(url, parentFile.getAbsolutePath());
             if (filePath.endsWith(".zip"))
             {
-                System.err.println("Unzipping to " + target.getAbsolutePath());
-                unzip(filePath, testPath.getAbsolutePath());
+                unzip(filePath, parentFile.getAbsolutePath());
             }
             return target.getAbsolutePath();
         }
@@ -44,6 +72,18 @@ public class TestUtility
             System.exit(1);
             return null;
         }
+    }
+
+    /**
+     * 保证 data/test/name 存在
+     *
+     * @param name
+     * @param url
+     * @return
+     */
+    public static String ensureTestData(String name, String url)
+    {
+        return ensureData(String.format("data/test/%s", name), url);
     }
 
     /**
@@ -83,8 +123,7 @@ public class TestUtility
             else
             {
                 // extracts file name from URL
-                fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
-                                             fileURL.length());
+                fileName = new File(httpConn.getURL().getPath()).getName();
             }
 
 //            System.out.println("Content-Type = " + contentType);
@@ -97,38 +136,47 @@ public class TestUtility
             String saveFilePath = savePath;
             if (new File(savePath).isDirectory())
                 saveFilePath = savePath + File.separator + fileName;
-            saveFilePath += ".downloading";
-
-            // opens an output stream to save into file
-            FileOutputStream outputStream = new FileOutputStream(saveFilePath);
-
-            int bytesRead;
-            byte[] buffer = new byte[4096];
-            long start = System.currentTimeMillis();
-            int progress_size = 0;
-            while ((bytesRead = inputStream.read(buffer)) != -1)
+            String realPath;
+            if (new File(saveFilePath).isFile())
             {
-                outputStream.write(buffer, 0, bytesRead);
-                long duration = (System.currentTimeMillis() - start) / 1000;
-                duration = Math.max(duration, 1);
-                progress_size += bytesRead;
-                int speed = (int) (progress_size / (1024 * duration));
-                float ratio = progress_size / (float) contentLength;
-                float percent = ratio * 100;
-                int eta = (int) (duration / ratio * (1 - ratio));
-                int minutes = eta / 60;
-                int seconds = eta % 60;
-
-                System.err.printf("\r%.2f%%, %d MB, %d KB/s, ETA %d min %d s", percent, progress_size / (1024 * 1024), speed, minutes, seconds);
+                System.err.printf("Use cached %s instead.\n", fileName);
+                realPath = saveFilePath;
             }
+            else
+            {
+                saveFilePath += ".downloading";
 
-            outputStream.close();
+                // opens an output stream to save into file
+                FileOutputStream outputStream = new FileOutputStream(saveFilePath);
+
+                int bytesRead;
+                byte[] buffer = new byte[4096];
+                long start = System.currentTimeMillis();
+                int progress_size = 0;
+                while ((bytesRead = inputStream.read(buffer)) != -1)
+                {
+                    outputStream.write(buffer, 0, bytesRead);
+                    long duration = (System.currentTimeMillis() - start) / 1000;
+                    duration = Math.max(duration, 1);
+                    progress_size += bytesRead;
+                    int speed = (int) (progress_size / (1024 * duration));
+                    float ratio = progress_size / (float) contentLength;
+                    float percent = ratio * 100;
+                    int eta = (int) (duration / ratio * (1 - ratio));
+                    int minutes = eta / 60;
+                    int seconds = eta % 60;
+
+                    System.err.printf("\r%.2f%%, %d MB, %d KB/s, ETA %d min %d s", percent, progress_size / (1024 * 1024), speed, minutes, seconds);
+                }
+                System.err.println();
+                outputStream.close();
+                realPath = saveFilePath.substring(0, saveFilePath.length() - ".downloading".length());
+                if (!new File(saveFilePath).renameTo(new File(realPath)))
+                    throw new IOException("Failed to move file");
+            }
             inputStream.close();
             httpConn.disconnect();
-            String realPath = saveFilePath.substring(0, saveFilePath.length() - ".downloading".length());
-            if (!new File(saveFilePath).renameTo(new File(realPath)))
-                throw new IOException("Failed to move file");
-            System.err.println();
+
             return realPath;
         }
         else
@@ -140,6 +188,7 @@ public class TestUtility
 
     private static void unzip(String zipFilePath, String destDir)
     {
+        System.err.println("Unzipping to " + destDir);
         File dir = new File(destDir);
         // create output directory if it doesn't exist
         if (!dir.exists()) dir.mkdirs();
@@ -179,11 +228,11 @@ public class TestUtility
             zis.closeEntry();
             zis.close();
             fis.close();
+            new File(zipFilePath).delete();
         }
         catch (IOException e)
         {
             e.printStackTrace();
         }
-
     }
 }
